requirements.txt
# ── Web framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1

# ── Async task queue ──────────────────────────────────────
celery==5.4.0
redis==5.0.5

# ── AI / ML APIs ──────────────────────────────────────────
openai==1.30.1
elevenlabs==1.2.0

# ── Audio / Video processing ──────────────────────────────
ffmpeg-python==0.2.0
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.1
pydub==0.25.1

# ── Speaker diarization ───────────────────────────────────
pyannote.audio==3.3.1
torch==2.3.0
torchaudio==2.3.0

# ── Storage ───────────────────────────────────────────────
boto3==1.34.110
aiofiles==23.2.1

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
structlog==24.2.0